seshat-node: Move the tasks into a separate module.
